refactor: padronizado as Views; criado css específicos para cada component; limpando index.css
