+++ src/components/ui/Filter/Filter.css
@@ -1,0 +1,73 @@
+.filter-container {
+  display: flex;
+  justify-content: space-between;
+}
+
+.filter-controls {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 24px;
+  align-items: flex-end;
+  width: 100%;
+}
+
+.control-group {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+
+.control-group label {
+  font-size: 13px;
+  font-weight: 500;
+  color: var(--muted);
+}
+
+.control-group input[type="month"],
+.input-with-prefix {
+  background-color: #f9fafb;
+  border: 1px solid var(--border-color);
+  border-radius: 8px;
+  height: 40px;
+  transition: border-color 0.2s, box-shadow 0.2s;
+}
+
+.control-group input[type="month"]:focus,
+.input-with-prefix:focus-within {
+  border-color: var(--success);
+  box-shadow: 0 0 0 2px rgba(34, 197, 94, 0.2);
+}
+
+.control-group input[type="month"] {
+  padding: 0 12px;
+  font-size: 14px;
+  color: var(--text-primary);
+}
+
+.input-with-prefix {
+  display: flex;
+  align-items: center;
+}
+
+.input-with-prefix span {
+  padding-left: 12px;
+  font-size: 14px;
+  color: var(--muted);
+}
+
+.input-with-prefix input {
+  border: none;
+  background: transparent;
+  outline: none;
+  padding: 0 12px 0 8px;
+  height: 100%;
+  width: 100%;
+  font-size: 14px;
+  color: var(--text-primary);
+}
+
+.control-group input[type="month"]:focus {
+  border-color: var(--success);
+  box-shadow: 0 0 0 2px rgba(34, 197, 94, 0.2);
+  outline: none;
+}
