+++ src/components/ui/NavBar/NavBar.jsx
@@ -1,4 +1,5 @@
 import { NavLink } from 'react-router-dom';
+import './NavBar.css';
 
 export default function NavBar() {
   return (
