+++ src/index.css
@@ -50,81 +50,6 @@ a {
   color: transparent;
 }
 
-/* Navbar styles */
-.navbar {
-  position: sticky;
-  top: 0;
-  z-index: 50;
-  background: rgba(255, 255, 255, 0.8);
-  backdrop-filter: blur(8px);
-  border-bottom: 1px solid #e5e7eb;
-}
-
-.navbar-inner {
-  max-width: 1280px;
-  margin-left: auto;
-  margin-right: auto;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 64px;
-  padding-left: 1rem;
-  padding-right: 1rem;
-}
-
-.navbar-brand {
-  display: flex;
-  align-items: center;
-  font-weight: 700;
-  font-size: 1.25rem;
-  color: var(--text);
-}
-
-.navbar-menu {
-  display: flex;
-  align-items: center;
-}
-
-.navbar-links {
-  display: flex;
-  gap: 1.5rem;
-  flex-wrap: wrap;
-  align-items: center;
-}
-
-.navbar-links a {
-  color: var(--text);
-  font-weight: 500;
-  transition: color 0.2s ease-in-out;
-}
-
-.navbar-links a:hover {
-  color: var(--accent);
-}
-
-.navbar-cta {
-  margin-left: 1rem;
-  display: flex;
-  align-items: center;
-}
-
-.navbar-links .navlink {
-  color: var(--text);
-  padding: 8px 0;
-  font-weight: 500;
-  text-decoration: none;
-  transition: color 0.2s ease-in-out;
-}
-
-.navbar-links .navlink:hover {
-  color: var(--accent);
-}
-
-.navbar-links .navlink.active {
-  color: var(--accent);
-  font-weight: 600;
-}
-
 /* Button styles */
 .button-primary {
   background: var(--accent);
@@ -154,33 +79,4 @@ a {
 
 .button-secondary:hover {
   background: #f5f5f5;
-}
-
-/* Footer styles */
-.footer {
-  background: #1f2937;
-  color: #9ca3af;
-  padding-top: 4rem;
-  padding-bottom: 2rem;
-}
-.footer a {
-  color: #9ca3af;
-  transition: color 0.2s;
-}
-.footer a:hover {
-  color: #ffffff;
-}
-.footer-title {
-  font-weight: 600;
-  margin-bottom: 1rem;
-  color: #ffffff;
-}
-.footer-grid {
-  display: grid;
-  gap: 2rem;
-}
-@media (min-width: 768px) {
-  .footer-grid {
-    grid-template-columns: repeat(4, 1fr);
-  }
 }
