+++ src/layouts/AppLayout.jsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import NavBar from '../components/ui/NavBar.jsx';
-import Footer from '../components/ui/Footer.jsx';
+import NavBar from '../components/ui/NavBar/NavBar.jsx';
+import Footer from '../components/ui/Footer/Footer.jsx';
 
 export default function AppLayout() {
   return (
